feat: add daily report and status
Daily report added on a schedule, will report to a CHANNEL_ID (set in
.env) daily at 09:00 server time.

Status updates now happen every 5 minutes and will update on boot. This
only tracks the reserved fuel usage for today.

Also added total sales volume to each report, and wrapped all responses
in an embed because it looks prettier.